backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1

# Pipecat - voice pipeline orchestration
pipecat-ai[silero,deepgram,elevenlabs,anthropic,livekit]==0.0.57

# LiveKit - WebRTC transport
livekit==0.17.3
livekit-api==0.7.1
